+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changes
 
-## 0.3.1-SNAPSHOT
+## 0.3.1
 
 - If `:test-refresh {:notify-command ["command" "arguments"]}` is in
   your project.clj then a summary message is passed into that command.